Switch from accessing the internal implementations (e.g. PersonResourceProvider) to a self-client, which allows things like e.g. graphql access for efficiently fetching a graph of specific elements of resources.
It's not trivial to directly invoke this without going via the client code. Check out the .search() method on PersonResourceProvider for example.

Add authorization model for use in access control policies. Using the raw FHIR model leads to extremely verbose policy definitions. It's also impossible since not all objects have a reference in the direction I need, for instance Person -> Practitioner is a direct reference, but PractitionerRole -> Practitioner is pointing the wrong way.

This commit doesn't work, the authz policies are not functional, but it gives some idea what's involved in implementing this.